requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
azure-core==1.34.0
azure-identity==1.23.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.4
dnspython==2.7.0
fastapi==0.115.14
frozenlist==1.7.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
microsoft-kiota-abstractions==1.9.4
microsoft-kiota-authentication-azure==1.9.4
microsoft-kiota-http==1.9.4
msal==1.32.3
msal-extensions==1.3.1
msgraph-core==1.3.5
multidict==6.6.2
opentelemetry-api==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
propcache==0.3.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pymongo==4.13.2
python-multipart==0.0.20
requests==2.32.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
std-uritemplate==2.0.5
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.35.0
yarl==1.20.1
zipp==3.23.0
